requirements.txt
anticaptchaofficial==1.0.58
attrs==23.1.0
Babel==2.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
numpy==1.26.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
py==1.11.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tkcalendar==1.6.1
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0

mysql-connector-python~=8.3.0